Support for pluggable CSS parsers. Currently supports css_parser and csspool.
csspool has a dependency on libcroco which is a gnome-based library not easy
to get installed on a shared host environment (e.g. Heroku). css_parser on
the other hand is pure ruby which is much better. But css_parser I don't
believe is as full featured of a CSS parser. So the preferred css parser should
be determined by the needs of the developer.

This is similar to how many library allow for pluggable XML parsers.

The big downside to this patch is the code for inline-style got ugly since
it was less code to make csspool look like css_parser than trying to make
css_parser look like csspool.