A big patch but it actually makes no functional change. Purely a refactor.
Adding the pluggable CSS Parser reduced the quality of the code due to
the fact that a selector was no longer an object but now an array. So the main
goal was to restore the selector as an object to make the code easier to
follow. In addition this gives us a place to hang additional functionality
related to the selector.

In addition to the above major change the following other bits were refactored.

* Made the main module less procedural and more OO by turning it into
  a class. This made it easier to hang auxilary that is re-used in the
  processing.
* Change the dom parsing to look at if it responds to just css vs. looking
  at the actual class. This allows any Nokogiri compatible API (including
  a Nokogiri document fragment).

There was only a slight change to the test. Other than that the tests run
as before but I believe it will be easier to fix some bugs now that the
code is closer (and hopefully better) to how it was before the pluggable
CSS parser was added.