Fixed bugs in pseudo code processing.
The pseduo code processing was too general. It was targeted at handling
things like a:hover but not so much for things like a:first-child.

In the first case we want to use the pseudo class inlined using the
experimental syntax IF the user opts in. But in the second case we simply
want to inline it on the first child.

The other bug was that it was searching for pseudo classes anywhere in
the selector instead of just at the end.

The end result of these bugs was that if I had something like:

table td:first-child a {....}

The styles would get zapped is pseudo class processing was disabled
and incorrectly applied if it was enabled.

We fix this by classifying the pseudo classes into two categories:

* state-based classes (like :hover, :active, etc)
* location-based classes (like :first-child)

Location-based classes are basically applied to the nodes and then treated
like a non-pseudo class. State based classes are skipped unless the pseudo
option is enabled (like before).

The only thing that is still not handled is state-based pseudo classes that
are not at the end of a selector. For example:

  a:hover img

I am guessing this means only apply the image styles if the a is hovered. But
in the context of inlining I am not sure what this means and doubt it is
supported. So I think the result is simply don't do it. :)