Remove required dependency on csspool.
The reason for adding the pluggable CSS parser was so this library can
run on platforms where compiling it might be a pain. By making it a
dependency we require that compilation even if it is not actually used.

This remove it as a dependency but keeps it as a development dependency
(since I assume it is needed for testing). Basically we are assuming CSSParser
as the default parser and only if they add CSSPool to their project will
they have to compile it.